examples/python/requirements.txt
Flask==1.0.2
requests-oauthlib==1.0.0
PyJWT==1.6.4
